(#9) Skip Windows tests
There is a bug in the Windows testing code that prevents tests from
passing even though the module is functional on Windows. The issue is
tracked in https://github.com/danielparks/puppetlabs-pipelines/issues/9.

There is another problem in rspec-puppet that breaks Windows testing,
but there is a functional work around:
https://github.com/rodjek/rspec-puppet/issues/662