CDPE-1583 fixes exec in unix.pp
Currently, agent install fails if the install_dir does not exist on the system. It gets to the Exec[mkdir -p ] and mkdir returns 1 because no dir is passed as an argument, the command on the resource is just 'mkdir -p'. This change passes install_dir to the command param to make this behave as presumably intended.